Begin filling out Search

+++ learning/learningAgent.go
@@ -2,6 +2,6 @@ package learning
 
 import "github.com/amirblum/SynergyAI/model"
 
-func LearnSynergy(world *model.World, team []model.Worker, task model.Task, score float32) {
+func LearnSynergy(world *model.World, team []model.Worker, task model.Task, score float64) {
 
 }
